+++ apps/apigateway/src/products/intelligence/commands/handler/create-query.handler.ts
@@ -1,11 +1,11 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs"
-import { CopilotRepository } from "../../copilot.repository"
+import { IntelligenceRepository } from "../../intelligence.repository"
 import { CreateQueryCommand } from "../impl/create-query.command"
 import { Types } from "mongoose"
 
 @CommandHandler(CreateQueryCommand)
 export class CreateQueryCommandHandler implements ICommandHandler<CreateQueryCommand> {
-  constructor(private readonly repository: CopilotRepository) { }
+  constructor(private readonly repository: IntelligenceRepository) { }
 
   async execute(command: CreateQueryCommand) {
     const { orgId, prompt, response } = command