+++ apps/apigateway/src/products/intelligence/intelligence.controller.ts
@@ -1,18 +1,18 @@
 import { Controller, Post, Body, BadRequestException, UseGuards, Request } from "@nestjs/common"
-import { CopilotService } from "./copilot.service"
+import { IntelligenceService } from "./intelligence.service"
 import { AIGenerationDto } from "./dto/ai-generate.dto"
 import { CredentialGuard } from "src/auth/credential.guard"
 import { ModRequest } from "src/auth/types/mod-request.interface"
 
-@Controller("products/copilot")
-export class CopilotController {
-  constructor(private readonly copilotService: CopilotService) { }
+@Controller("products/intelligence")
+export class IntelligenceController {
+  constructor(private readonly intelligenceService: IntelligenceService) { }
 
   @UseGuards(CredentialGuard)
   @Post("generate")
   async generateRecommendation(@Request() request: ModRequest, @Body() aiGenerationDto: AIGenerationDto) {
     try {
-      return await this.copilotService.generateRecommendation(request.user.orgId, aiGenerationDto.prompt, aiGenerationDto.temperature, aiGenerationDto.topP, aiGenerationDto.topK)
+      return await this.intelligenceService.generateRecommendation(request.user.orgId, aiGenerationDto.prompt, aiGenerationDto.temperature, aiGenerationDto.topP, aiGenerationDto.topK)
     }
 
     catch (error) {