+++ apps/apigateway/src/products/intelligence/intelligence.repository.ts
@@ -6,8 +6,8 @@ import { Model } from "mongoose"
 import { BaseRepository } from "src/shared/database/database.repository"
 
 @Injectable()
-export class CopilotRepository extends BaseRepository<Query> {
-  constructor(@InjectModel(Query.name, DbConnectionMap.Copilot) private queryModel: Model<Query>) {
+export class IntelligenceRepository extends BaseRepository<Query> {
+  constructor(@InjectModel(Query.name, DbConnectionMap.Intelligence) private queryModel: Model<Query>) {
     super(queryModel)
   }
 }