+++ apps/apigateway/src/products/products.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common"
-import { CopilotModule } from "./copilot/copilot.module"
+import { IntelligenceModule } from "./intelligence/intelligence.module"
 import { DatamarketplaceModule } from "./datamarketplace/datamarketplace.module"
 import { HttpnosqlModule } from "./httpnosql/httpnosql.module"
 import { WebAnalyticsModule } from "./webanalytics/webanalytics.module"
@@ -7,7 +7,7 @@ import { IdentityModule } from "./identity/identity.module"
 
 @Module({
   imports: [
-    CopilotModule,
+    IntelligenceModule,
     DatamarketplaceModule,
     HttpnosqlModule,
     WebAnalyticsModule,