+++ apps/apigateway/src/shared/utils/db-connection.map.ts
@@ -1,8 +1,8 @@
 export enum DbConnectionMap {
   Core = "core",
-  Copilot = "copilot",
   DataMarketplace = "datamarketplace",
   HttpNoSql = "httpnosql",
   Identity = "identity",
+  Intelligence = "intelligence",
   WebAnalytics = "webanalytics"
 }